+++ src/main/java/com/transactionapi/dto/TradeHistoryResponse.java
@@ -20,7 +20,7 @@ public record TradeHistoryResponse(
         AssetType assetType,
         Currency currency,
         TradeDirection direction,
-        int quantity,
+        BigDecimal quantity,
         BigDecimal entryPrice,
         BigDecimal exitPrice,
         BigDecimal fees,
